+++ lib/v2/screens/explore_screens/vote_screens/proposal_details/proposal_details_screen.dart
@@ -25,7 +25,7 @@ class ProposalDetailsScreen extends StatelessWidget {
       body: CustomScrollView(
         slivers: [
           SliverAppBar(
-            expandedHeight: 250,
+            expandedHeight: 280,
             flexibleSpace: Stack(
               fit: StackFit.expand,
               children: [
